@@ -30,17 +30,25 @@
     "@nuxtjs/color-mode": "^3.3.2",
     "@pinia-plugin-persistedstate/nuxt": "^1.2.0",
     "@searchkit/instantsearch-client": "^4.13.0",
+    "@testing-library/vue": "^8.1.0",
     "@types/jsonwebtoken": "^9.0.5",
     "@types/lodash": "^4.17.0",
     "@types/mime": "^4.0.0",
     "@types/pg": "^8.10.9",
+    "@typescript-eslint/eslint-plugin": "^7.16.1",
+    "@typescript-eslint/parser": "^7.16.1",
     "@typescript-eslint/typescript-estree": "^7.14.1",
-    "@vueuse/core": "^10.7.2",
+    "@vitejs/plugin-vue": "^5.0.5",
+    "@vitest/ui": "^2.0.3",
+    "@vueuse/core": "^10.11.0",
     "@vueuse/nuxt": "^10.7.2",
     "caniuse-lite": "^1.0.30001632",
+    "cypress": "^13.13.1",
     "daisyui": "^4.7.2",
     "diff-match-patch-ts": "^0.5.0",
     "elasticsearch": "^16.7.3",
+    "esbuild": "^0.23.0",
+    "eslint-plugin-cypress": "^3.3.0",
     "faker": "^6.6.6",
     "highlight.js": "^11.9.0",
     "lodash": "^4.17.21",
@@ -52,7 +60,9 @@
     "searchkit": "^4.12.0",
     "theme-change": "^2.5.0",
     "tsx": "^4.7.2",
+    "unplugin-auto-import": "^0.18.0",
     "vite": "^5.3.1",
+    "vite-plugin-commonjs": "^0.10.1",
     "vue": "^3.4.30",
     "vue-diff": "^1.2.4",
     "vue-instantsearch": "^4.17.5",
@@ -67,6 +77,7 @@
     "@cloudflare/workers-types": "^4.20240620.0",
     "@nuxt/devtools": "^1.3.6",
     "@nuxt/eslint-config": "^0.2.0",
+    "@nuxt/test-utils": "^3.13.1",
     "@nuxt/types": "^2.17.3",
     "@nuxt/typescript-build": "^3.0.2",
     "@nuxtjs/eslint-module": "^4.1.0",
@@ -74,13 +85,14 @@
     "@nuxtjs/strapi": "^1.11.0",
     "@nuxtjs/tailwindcss": "^6.10.1",
     "@pinia/nuxt": "^0.5.1",
-    "@types/node": "^20.11.0",
+    "@types/node": "^20.14.11",
+    "@vue/test-utils": "^2.4.6",
     "eslint": "^8.54.0",
     "nitro-cloudflare-dev": "^0.1.4",
     "nuxt-icon": "^0.6.8",
     "sass": "^1.69.5",
     "ts-node": "^10.9.2",
-    "typescript": "^5.3.3",
+    "typescript": "5.5.3",
     "vue": "^3.4.30",
     "vue-router": "^4.2.5",
     "wrangler": "^3.61.0"
