@@ -45,7 +45,7 @@ const response = null;
       <div>
         <div
           role="alert"
-          class="alert"
+          class="alert bg-neutral"
         >
           <svg
             xmlns="http://www.w3.org/2000/svg"
@@ -64,43 +64,21 @@ const response = null;
         </div>
       </div>
       <article class="container">
-        <h1>Garlic bread with cheese: What the science tells us</h1>
-        <p>
-          For years parents have espoused the health benefits of eating garlic bread with cheese to their
-          children, with the food earning such an iconic status in our culture that kids will often dress
-          up as warm, cheesy loaf for Halloween.
-        </p>
-        <p>
-          But a recent study shows that the celebrated appetizer may be linked to a series of rabies cases
-          springing up around the country.
-        </p>
+        <div
+          id="c7273"
+          class="frame frame-default frame-type-text frame-layout-0"
+        >
+          <p>Im Projekt <em>AV-EFI Automatisiertes Verbundsystem für audiovisuelle Bestände über einheitliche Filmidentifikatoren</em> – Eigenschreibweise mit Projektstart <strong>AVefi</strong> – wird ein institutionsübergreifendes, webbasiertes System entwickelt und erprobt, mit Persistent Identifier (PID) in zentraler Rolle. Dieses soll der Identifikation und Verknüpfung von Filmen, ihren Fassungen und Datenobjekten aus heterogenen Beständen dienen.</p>
+          <h3>Projektgegenstand</h3>
+          <p>Die Vernetzung standardisierter wissenschaftlicher Metadaten über Institutionen hinweg ist ein übergeordnetes gesellschaftliches Ziel. Es ermöglicht effizientere Strukturen, weitere Innovationen und knüpft damit an die Tradition der großen Verbünde im Archiv- und Bibliotheksbereich an. Zu diesem übergeordneten Ziel leistet das Projekt <strong>AVefi </strong>einen neuartigen Beitrag auf dem Gebiet audiovisueller Bestände und ihrer Nachweise für die wissenschaftliche Forschung und das Kulturerbe. Es werden insbesondere Film- und Medienwissenschaften, aber auch die interdisziplinären Geistes- und Kulturwissenschaften, vor allem die Digital Humanities, adressiert.</p>
+          <p>Im Projekt wird entwickelt und erprobt, wie ein institutionsübergreifendes, webbasiertes System, in dem Persistent Identifier (PID) eine zentrale Rolle spielen, als Forschungsinfrastruktur zur Identifikation von AV-Werken, den dazugehörenden Fassungen und Datenobjekten dient und wie sie miteinander verknüpft werden können. Das PID-System mit den einheitlichen Filmidentifikatoren (EFI) übernimmt die Funktion einer zentralen Datenbank, führt die existierenden Metadaten der unterschiedlichen Fassungen und Materialien aus den originalen Datenbanken verschiedener Institutionen strukturiert und möglichst automatisiert zusammen und sichert damit die freie und langfristige Verfügbarkeit dieser Informationen für die Öffentlichkeit. Dabei ist angedacht, die Disambiguierung und die Zuordnung der Metadaten verschiedener Institutionen durch automatische Synchronisations- und Abgleichsroutinen zu unterstützen. Das zu entwickelnde Verbundsystem einschließlich einer benutzungsfreundlichen Recherche- und Redaktionsoberfläche, soll das bereits verfügbare “Ökosystem” (filmischer) Identifikatoren und (Norm-)Daten (z.B. Filmportal, EIDR, GND, NFDI4Culture) sinnvoll ergänzen und sich in dieses als originäres Angebot mit starkem Bestandsbezug einfügen.</p>
+          <h3>Projektziele</h3>
+          <p>Das initiale zweijährige Projekt dient in enger Zusammenarbeit mit Praxispartnern und Vertreter*innen von Forschungscommunities nicht nur der beispielhaften, modularen Entwicklung und Erprobung des Systems auf technischer Ebene. Es gilt nicht zuletzt im intensiven Austausch, insbesondere bzgl. des Metadatenmanagements gemeinsam detaillierte Anforderungen, Regeln und Workflows im Verbund mit dem Ziel einer anknüpfüngsfähigen, dauerhaften Gestalt zu erarbeiten.&nbsp;</p>
+          <p>Langfristig soll das System dem Nachweis und der Verknüpfung von AV-Beständen jeglicher Herkunft und Gattung dienen. Mit dem Projekt wird zudem eine Lücke bearbeitet: mit besonderem Blick auf Unterrichts-, Lehr und Wissenschaftsfilme sowie auf ephemeres Filmmaterial oder Amateurfilme sollen diese Gattungen erstmals in einem Verbundsystem als “kleiner” Film” neben dem “großen” Film der Forschung zur Verfügung stehen. Durch eine niedrigschwellige, automatisierte, ressourcenschonende Technologie, wird insbesondere auch kleineren Institutionen die Möglichkeit der Projektteilnahme und -teilhabe gegeben.</p>
+          <h3>Akteure</h3>
+          <p>Verbundprojektpartner sind: die TIB - Leibniz-Informationszentrum für Technik und Naturwissenschaften, die Stiftung Deutsche Kinemathek – Museum für Film und Fernsehen (SDK),&nbsp;Gesellschaft für wissenschaftliche Datenverarbeitung mbH Göttingen (GWDG) und das&nbsp;Filmmuseum der Landeshauptstadt Düsseldorf (FMDU). Diese werden von zehn Praxispartnern in verschiedenen Rollen begleitet und unterstützt</p>
+        </div>
       </article>
-      <NuxtLayout name="partial-layout-3-2">
-        <template #title>
-          FAQs
-        </template>
-        <template #containerLeft>
-          Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet.   
-
-          Duis autem vel eum iriure dolor in hendrerit in vulputate velit esse molestie consequat, vel illum dolore eu feugiat nulla facilisis at vero eros et accumsan et iusto odio dignissim qui blandit praesent luptatum zzril delenit augue duis dolore te feugait nulla facilisi. Lorem ipsum dolor sit amet, consectetuer adipiscing elit, sed diam nonummy nibh euismod tincidunt ut laoreet dolore magna aliquam erat volutpat.   
-
-          Ut wisi enim ad minim veniam, quis nostrud exerci tation ullamcorper suscipit lobortis nisl ut aliquip ex ea commodo consequat. Duis autem vel eum iriure dolor in hendrerit in vulputate velit esse molestie consequat, vel illum dolore eu feugiat nulla facilisis at vero eros et accumsan et iusto odio dignissim qui blandit praesent luptatum zzril delenit augue duis dolore te feugait nulla facilisi.   
-
-          Nam liber tempor cum soluta nobis eleifend option congue nihil imperdiet doming id quod mazim placerat facer possim assum. Lorem ipsum dolor sit amet, consectetuer adipiscing elit, sed diam nonummy nibh euismod tincidunt ut laoreet dolore magna aliquam erat volutpat. Ut wisi enim ad minim veniam, quis nostrud exerci tation ullamcorper suscipit lobortis nisl ut aliquip ex ea commodo consequat.   
-
-          Duis autem vel eum iriure dolor in hendrerit in vulputate velit esse molestie consequat, vel illum dolore eu feugiat nulla facilisis.   
-
-          At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, At accusam aliquyam diam diam dolore dolores duo eirmod eos erat, et nonumy sed tempor et et invidunt justo labore Stet clita ea et gubergren, kasd magna no rebum. sanctus sea sed takimata ut vero voluptua. est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat.   
-
-          Consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita
-        </template>
-        -->
-        <template #containerRight>
-          Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit amet.
-          Duis autem vel eum iriure dolor in hendrerit in vulputate velit esse molestie consequat, vel illum dolore eu feugiat nulla facilisis at vero eros et accumsan et iusto odio dignissim qui blandit praesent luptatum zzril delenit augue duis dolore te feugait nulla facilisi. Lorem ipsum dolor sit amet, consectetuer adipiscing elit, sed diam nonummy nibh euismod tincidunt ut laoreet dolore magna aliquam erat volutpat.   
-          Ut wisi enim ad minim veniam, quis nostrud exerci tation ullamcorper suscipit lobortis nisl ut aliquip ex ea commodo consequat. Duis autem vel eum iriure dolor in hendrerit in vulputate velit esse molestie consequat, vel illum dolore      
-        </template>
-      </NuxtLayout>
     </div>
   </div>
 </template>
