@@ -3,18 +3,9 @@ definePageMeta({
     //middleware: 'auth'
 });
 const route = useRoute();
-let arr = [];
-if(route.query.compare) {
-    arr = route.query.compare;
-}
-const objectListStore = useObjectListStore();
-
-const navigateToComparisonAltern = () => {
-    const objectIds: number[] = objectListStore.getObjectIds;
-    if(objectIds.length == 2) {
-        navigateTo(`/protected/compare_altern?compare=[${objectIds[0]},${objectIds[1]}]`);
-    }
-};
+const items = new Array();
+items[0] = route.query.prev;
+items[1] = route.query.next;
 
 
 </script>
@@ -33,15 +24,8 @@ const navigateToComparisonAltern = () => {
       </div>
     </div>
     <div class="container">
-      <button
-        title="Navigate to Alternate View"
-        class="btn btn-primary btn-ghost hidden"
-        @click="navigateToComparisonAltern"
-      >
-        Alternativ
-      </button>
-      <div v-if="arr">
-        <GlobalCompareView :items="objectListStore.getObjectIds" />
+      <div v-if="items">
+        <GlobalCompareViewRaw :items="items" />
       </div>
       <div v-else>
         no data to compare
