@@ -1,15 +1,10 @@
 export const config = {
+    //Currently not used
     connection: {
-        //host: 'https://commerce-demo.es.us-east4.gcp.elastic-cloud.com:9243',
-        host: 'https://technische-informati-8716937806.eu-central-1.bonsaisearch.net/kibana_sample_data_ecommerce',
-        // if you are authenticating with api key
-        // https://www.searchkit.co/docs/guides/setup-elasticsearch#connecting-with-api-key
-        //apiKey: 'a2Rha1VJTUJMcGU4ajA3Tm9fZ0Y6MjAzX2pLbURTXy1hNm9SUGZGRlhJdw=='
-        // if you are authenticating with username/password
-        // https://www.searchkit.co/docs/guides/setup-elasticsearch#connecting-with-usernamepassword
+        host: process.env.ELASTIC_BONSAI_HOST,
         auth: {
-            username: "58xyjh39p8",
-            password: "1wyuvqg1ax"
+            username: process.env.ELASTIC_BONSAI_USERNAME,
+            password: process.env.ELASTIC_BONSAI_PASSW
         },
     },
     search_settings: {
